requirements.txt
python-dotenv
requests
groq
fastapi
uvicorn